decimal figures in price bug fixed

+++ scripts/add_to_monIntExp.sql
@@ -7,7 +7,7 @@ exec dbo.z_NewChID 't_MonIntExp', @NewChID OUTPUT
 
 
 
-declare  @CRID SMALLINT, @OurID INT
+declare  @CRID INT, @OurID INT
 
 select  @CRID=c.CRID, @OurID = r.OurID
 from r_Crs c
